[Runtime] "Set Volume to" "Change Volume by" and "Volume" blocks added

+++ source/scratch/sprite.hpp
@@ -378,7 +378,7 @@ class Sprite {
     bool isDeleted = false;
     int currentCostume;
     std::string lastCostumeId = "";
-    int volume;
+    float volume;
     double xPosition;
     double yPosition;
     int rotationCenterX;
